Move deprecated math/cylindrical to matlab/util
