+++ drake/math/BUILD
@@ -7,7 +7,6 @@ cc_library(
         "autodiff.h",
         "axis_angle.h",
         "cross_product.h",
-        "cylindrical.h",
         "gradient.h",
         "quaternion.h",
         "random_rotation.h",
@@ -19,7 +18,6 @@ cc_library(
         "autodiff_gradient.cc",
         "axis_angle.cc",
         "cross_product.cc",
-        "cylindrical.cc",
         "gradient.cc",
         "quaternion.cc",
         "random_rotation.cc",
